% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_data.R
\name{compose_data}
\alias{compose_data}
\title{Compose data from fixed and random effects, with normally distributed errors}
\usage{
compose_data(design_args, n_subj = NULL, fixed = NULL, subj_rmx = NULL,
  item_rmx = NULL, err_var = 1, verbose = FALSE)
}
\arguments{
\item{design_args}{List containing information about the experiment
design; see \code{\link{stim_lists}}}

\item{fixed}{vector of fixed effects}

\item{subj_rmx}{matrix of by-subject random effects}

\item{item_rmx}{matrix of by-item random effects}

\item{err_var}{error variance}

\item{verbose}{give debugging info (default = \code{FALSE})}
}
\value{
a data frame containing simulated data
}
\description{
Compose data from fixed and random effects, with normally distributed errors
}

