% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{contr.deviation}
\alias{contr.deviation}
\title{Make Matrix for Deviation-Coded Predictors}
\usage{
contr.deviation(n)
}
\arguments{
\item{n}{A vector of levels for a factor, or the number of levels.}
}
\description{
Return a matrix of deviation-coded predictors (analogous to
\code{\link{contr.sum}}).
}

