% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{fitnocorr.mcmc}
\alias{fitnocorr.mcmc}
\title{Run mixed-effects model (\code{\link{lmer}}) with MCMC p-value}
\usage{
fitnocorr.mcmc(mcr.data, wsbi = FALSE, nmcmc = 10000)
}
\arguments{
\item{mcr.data}{A dataframe formatted as described in \code{\link{mkDf}}.}

\item{wsbi}{Whether the design is between-items (TRUE) or within-items
(FALSE).}

\item{nmcmc}{Number of Markov-Chain Monte Carlo simulations (default =
10000).}
}
\value{
A single element vector with the mcmc p-value.
}
\description{
Uses \code{\link{lmer}} to fit a random-intercepts only model, and then uses
\code{\link{mcmcsamp}} to obtain p-values.
}
\details{
If the model does not converge, returns \code{NA}.  The MCMC procedure is
based on Baayen's \code{pvals.fnc} in package \code{languageR}.
}
\note{
This function ONLY fits models with independent random intercepts and
slopes.
}
\examples{
nmc <- 10
pmx <- cbind(randParams(genParamRanges(), nmc, 1001), seed=mkSeeds(nmc, 1001))

# between-items dataset
x.bi <- mkDf(nsubj=24, nitem=24, mcr.params=pmx[1,], wsbi=TRUE)

# NB: small number of MCMC runs so that the example runs quickly
# increase the number of runs for stable results
fitnocorr.mcmc(x.bi, wsbi=TRUE, nmcmc=1000)
}
\seealso{
\code{\link{fitnocorr}}
}

