% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{formVCovMx}
\alias{formVCovMx}
\title{Form a symmetric variance-covariance matrix}
\usage{
formVCovMx(vars, corrs)
}
\arguments{
\item{vars}{vector of variances}

\item{corrs}{vector of correlations}
}
\value{
A symmetric variance-covariance matrix
}
\description{
Takes a vector of variances and correlations, calculates
covariances and forms a symmetric matrix (for use in generating
data from a multivariate normal distribution)
}
\examples{
pars <- findValidVCovParams(vars=list(A=c(1,3),B=c(1,3), C=c(0,2)),
                            corrs=list(AB=c(-.5,.5),AC=c(-.5,.5),BC=c(-.5,.5)))

formVCovMx(pars[1:3], pars[4:6])
}

