\name{genParamRanges}
\alias{genParamRanges}
\title{Data-generating parameter ranges}
\usage{
  genParamRanges()
}
\value{
  A list containing ranges for the following parameters,
  default values in parentheses:
}
\description{
  Generates parameter ranges for one-factor design.  These
  parameters are used for generating datasets.
}
\seealso{
  \code{\link{genParamRanges.facMixedAB}},
  \code{\link{randParams}}, \code{\link{mkDf}}
}

