% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{loessPred}
\alias{loessPred}
\title{Fit a loess model to the p-value data for later use in generating heatmaps.}
\usage{
loessPred(pval, paramX, paramY, minx = 0.01, maxx = 2.99, miny = 0.01,
  maxy = 2.99, svec = rep(TRUE, length(paramX)), span = 0.9)
}
\arguments{
\item{pval}{vector of p-values from simulated runs}

\item{paramX}{population parameter to condition \code{z} on (e.g., a column
from a matrix created using \code{\link{createParamMx}}).  This will appear
on the x-axis of the heatmap.}

\item{paramY}{same as \code{paramX}, but for the Y-axis}

\item{minx}{minimum x value}

\item{maxx}{maximum x value}

\item{miny}{minimum y value}

\item{maxy}{maximum y value}

\item{svec}{a logical vector to select out only certain rows from
\code{pval}, \code{paramX} and \code{paramY}}

\item{span}{a value to be passed along to the \code{\link{loess}} function
that determines the smoothing level}
}
\value{
A fitted \code{\link{loess}} model.
}
\description{
Given a vector of p-values and two population parameters, this will fit a
\code{\link{loess}} model to the data, with the model of the form z ~ x * y,
where z is the predicted p-value.  This can later be used to make a heatmap,
and is separated from the \code{\link{heatmap2}} function because of the
long amount of time it can take for this function to run.
}
\examples{
pmx <- createParamMx(100)

result <- apply(pmx, 1, function(x) {fitanova(mkDf(24,24,mcr.params=x,FALSE),FALSE)})
lpred <- loessPred(pval=result["pmax",],
                paramX=pmx[,"t11"], paramY=pmx[,"w11"])

heatmap2(lpred, "t11", "w11", "F1+F2")
}
\seealso{
\code{\link{loess}}, \code{\link{heatmap2}}
}

