\name{mkDf}
\alias{mkDf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Make a dataframe with simulated data given a set of population parameters
}
\description{
Generates data for a simulated experiment given population parameters and information
about the design.
}
\usage{
mkDf(nsubj=24, nitem=24, wsbi=FALSE, mcr.params=createParamMx(1,firstseed=sample(1:1000000,1))[1,],
     missMeth = 'random', rigen = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{nsubj}{
number of subjects in the experiment
}
  \item{nitem}{
number of items in the experiment
}
  \item{mcr.params}{
population parameters (typically a row from the matrix generated by \code{\link{createParamMx}}
}
  \item{wsbi}{
whether the design is between-items (TRUE) or within-items (FALSE)
}
\item{missMeth}{
method used for generating missing data (default is 'random';
alternatives are 'none', 'randomBig', 'bysubj', 'bycond', bysubjcond'; see Details)
}
\item{rigen}{
  use a 'random-intercepts-only' generative model (default FALSE)
  }
}
\details{
Note that for between-items designs, the variable \code{w11} in the
parameter vector is simply ignored.  The default value for
\code{missMeth} of 'random' will randomly remove between 0-5\% of the
observations.  The option 'randomBig' will randomly remove 10-80\%;
'bysubj' will randomly remove 10-80\% of each subject's data; 'bycond'
will randomly remove 10-80\% of each condition's data; and 'bysubjcond'
will randomly remove 10=80\% of each subject/condition combination's data.
}
\value{
  A dataframe, with fields:
  \item{SubjID}{a factor, identifying subject number}
  \item{ItemID}{a factor, identifying item number}
  \item{Cond}{treatment condition, deviation coded (-.5, .5)}
  \item{Resp}{response variable}
}
%\references{
%% ~put references to the literature/web site here ~
%}
%\author{
%%  ~~who you are~~
%}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{createParamMx}}
}
\examples{
pmx <- createParamMx(10)

x.df <- mkDf(nsubj=24, nitem=24, mcr.params=pmx[1,], wsbi=FALSE)

## set rate of missing observations manually
pmx[,"miss"] <- 0.4
x.df2 <- mkDf(nsubj=24, nitem=24, mcr.params=pmx[1,], wsbi=FALSE)

## by-condition missing observation rates
pmx[,"pMin"] <- 0.1
pmx[,"pMax"] <- 0.8
x.df3 <- mkDf(nsubj=24, nitem=24, mcr.params=pmx[1,], wsbi=FALSE,missMeth="bycond")

## by-subject missing observation rates
x.df4 <- mkDf(nsubj=24, nitem=24, mcr.params=pmx[1,], wsbi=FALSE,missMeth="bysubj")

## by-subject/condition pair missing observation rates
x.df5 <- mkDf(nsubj=24, nitem=24, mcr.params=pmx[1,], wsbi=FALSE,missMeth="bysubjcond")


}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
