% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{randParams}
\alias{randParams}
\title{Generate population parameters from data-generating parameters}
\usage{
randParams(plist, nmc = 1, firstseed = NULL)
}
\arguments{
\item{plist}{Named list of parameters (e.g., from
\code{\link{genParamRanges}}), with each element of the list a
vector that defines the parameter's range (min, max).  If the
vector has only a single element, that parameter is treated as
constant.}

\item{nmc}{Number of populations to create}

\item{firstseed}{Seed the random number generator before generating
the populations (unless \code{NULL}).}
}
\value{
A matrix where each row has the parameter values for a population
}
\description{
Takes a list of ranges for data-generating parameters and randomly
instantiates parameters for a series of populations, drawing each
value from a uniform distribution.
}
\examples{
randParams(genParamRanges(), 2, 1001)
randParams(genParamRanges(), 2, 1001) # same result as above
randParams(genParamRanges(), 10) # different

}

