% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_norm}
\alias{sim_norm}
\title{Simulate data with normally distributed errors}
\usage{
sim_norm(ivs, between_subj = c(), between_item = c(), n_subj = NULL,
  n_item = NULL, n_rep = 1, fixed = NULL, subj_rmx = NULL,
  item_rmx = NULL, err_var = 1, verbose = FALSE)
}
\arguments{
\item{ivs}{named list of independent variables (IVs) with each list
element a vector giving the levels of that IV, or a single number
stating the number of desired levels}

\item{between_subj}{character vector with names of IVs whose
levels are administered between subjects}

\item{between_item}{charater vector with names of IVs whose levels
are administered between items}

\item{n_item}{desired number of stimulus items; if \code{NULL},
will generate lists with the minimum possible number}

\item{n_rep}{number of repetitions of each stimulus item for each
participant (default 1)}

\item{fixed}{vector of fixed effects}

\item{subj_rmx}{matrix of by-subject random effects}

\item{item_rmx}{matrix of by-item random effects}

\item{err_var}{error variance}

\item{verbose}{give debugging info (default = \code{FALSE})}
}
\value{
a data frame containing simulated data
}
\description{
Simulate data with normally distributed errors
}

