% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_data.R
\name{term_names}
\alias{term_names}
\title{Derive names for a given design}
\usage{
term_names(ivs, between_subj = c(), between_item = c())
}
\arguments{
\item{ivs}{named list of independent variables (IVs) with each list
element a vector giving the levels of that IV, or a single number
stating the number of desired levels}

\item{between_subj}{character vector with names of IVs whose
levels are administered between subjects}

\item{between_item}{charater vector with names of IVs whose levels
are administered between items}
}
\value{
A vector of term names
}
\description{
Derive names of predictor terms for a given design
}
\seealso{
\code{\link{stim_lists}}
}

