% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/general.R
\name{contr.deviation}
\alias{contr.deviation}
\title{Make Matrix for Deviation-Coded Predictors}
\usage{
contr.deviation(n)
}
\arguments{
\item{n}{A vector of levels for a factor, or the number of levels.}
}
\description{
Return a matrix of deviation-coded predictors (analogous to
\code{\link{contr.sum}}).
}

