% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefactor.R
\name{fitlmer.mcmc}
\alias{fitlmer.mcmc}
\title{Run mixed-effects model (\code{\link{lmer}}) with MCMC p-value}
\usage{
fitlmer.mcmc(mcr.data, wsbi, nmcmc = 10000)
}
\arguments{
\item{mcr.data}{A dataframe formatted as described in \code{\link{mkDf}}.}

\item{wsbi}{Whether the design is between-items (TRUE) or within-items
(FALSE); has no effect because the model is random-intercepts only, but was
included for consistency with \code{\link{fitlmer}} and
\code{\link{fitanova}}.}

\item{nmcmc}{Number of Markov-Chain Monte Carlo simulations (default =
10000).}
}
\description{
Uses \code{\link{lmer}} to fit a random-intercepts only model, and then uses
\code{\link{mcmcsamp}} to obtain p-values (one-factor design only).
}
\details{
Tries to fit the most complex model requested, and if that fails to
converge, then tries progressively simpler models, before calculating mcmc
p-value.  If no model converges, returns NAs.  The MCMC procedure is based
on Baayen's \code{pvals.fnc} in package \code{languageR}.
}
\note{
This function no longer works and will throw an error, as the \code{mcmcsamp} function was removed starting with \code{lme4} package 1.0; see \code{\link[lme4]{mcmcsamp}} for details.
}
\seealso{
\code{\link{fitlmer}}
}
