% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefactor.R
\name{getParamRanges}
\alias{getParamRanges}
\title{Data-generating parameter ranges for one-factor design}
\usage{
getParamRanges()
}
\value{
An environment object, containing the following variables:
\item{evar.range}{range of error variance (default [0,3])}
\item{icept.range}{range of intercept (default [-3, 3])}
\item{pmissing.range}{range of proportion missing obs (default [0,
.05])}
\item{r01.item.range}{range of by-subject intercept/slope correlation
(default [-.8, .8])}
\item{r01.subj.range}{range of by-item intercept/slope correlation
(default [-.8, .8])}
\item{slope}{treatment effect, when h0 TRUE, 0; when h0 FALSE, .8}
\item{t00.range}{range of by-subject random intercept variance (default
[0, 3]}
\item{t11.range}{range of by-subject random slope variance (default [0,
3])}
\item{w00.range}{range of by-item random intercept variance (default
[0, 3])}
\item{w11.range}{range of by-item random slope variance (default [0,
3])}
}
\description{
Get the default list of parameter ranges used in the Monte Carlo
simulations for one-factor design.  These parameter ranges can then
be modified by the user.
}
\examples{

# display all the parameter ranges
print(as.list(getParamRanges()))

# how to alter one of the values
p.env <- getParamRanges()
p.env$pmissing.range <- c(0, .25)
assign("pmissing.range", c(0, .25), env=p.env) # alternate way
print(as.list(p.env))

}
\seealso{
\code{\link{createParamMx}}
}
