% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{mkDf}
\alias{mkDf}
\title{Make a dataframe with simulated data given a set of population parameters}
\usage{
mkDf(nsubj = 24, nitem = 24, wsbi = FALSE,
  mcr.params = randParams(genParamRanges(), 1)[1, ], missMeth = "random",
  rigen = FALSE)
}
\arguments{
\item{nsubj}{number of subjects in the experiment}

\item{nitem}{number of items in the experiment}

\item{mcr.params}{population parameters (typically a row from the matrix
generated by \code{\link{createParamMx}}}

\item{wsbi}{whether the design is between-items (TRUE) or within-items
(FALSE)}

\item{missMeth}{method used for generating missing data (default is
'random'; alternatives are 'none', 'randomBig', 'bysubj', 'bycond',
bysubjcond'; see Details)}

\item{rigen}{use a 'random-intercepts-only' generative model (default FALSE)}
}
\value{
A dataframe, with fields:
\item{SubjID}{a factor, identifying subject number}
\item{ItemID}{a factor, identifying item number}
\item{Cond}{treatment condition, deviation coded (-.5, .5)}
\item{Resp}{response variable}
}
\description{
Generates data for a simulated experiment given population parameters and
information about the design (one-factor design).
}
\details{
Note that for between-items designs, the variable \code{w11} in the
parameter vector is simply ignored.  The default value for \code{missMeth}
of 'random' will randomly remove between 0-5\% of the observations.  The
option 'randomBig' will randomly remove 10-80\%; 'bysubj' will randomly
remove 10-80\% of each subject's data; 'bycond' will randomly remove 10-80\%
of each condition's data; and 'bysubjcond' will randomly remove 10=80\% of
each subject/condition combination's data.
}
\examples{
nmc <- 10
pmx <- cbind(randParams(genParamRanges(), nmc, 1001), seed=mkSeeds(nmc, 1001))

x.df <- mkDf(nsubj=24, nitem=24, mcr.params=pmx[1,], wsbi=FALSE)

## set rate of missing observations manually
pmx[,"miss"] <- 0.4
x.df2 <- mkDf(nsubj=24, nitem=24, mcr.params=pmx[1,], wsbi=FALSE)

## by-condition missing observation rates
pmx[,"pMin"] <- 0.1
pmx[,"pMax"] <- 0.8
x.df3 <- mkDf(nsubj=24, nitem=24, mcr.params=pmx[1,], wsbi=FALSE,missMeth="bycond")

## by-subject missing observation rates
x.df4 <- mkDf(nsubj=24, nitem=24, mcr.params=pmx[1,], wsbi=FALSE,missMeth="bysubj")

## by-subject/condition pair missing observation rates
x.df5 <- mkDf(nsubj=24, nitem=24, mcr.params=pmx[1,], wsbi=FALSE,missMeth="bysubjcond")
}
\seealso{
\code{\link{genParamRanges}}, \code{\link{mkDf.facMixedAB}}
}

