\name{simgen-package}
\alias{simgen-package}
\alias{simgen}
\docType{package}
\title{
Monte Carlo Simulation of Simultaneous Generalization to Subjects and Items
}
\description{
Routines for running Monte Carlo simulations testing the
generalizability of various ways of analyzing datasets with crossed
random effects.  This accompanies a paper by Barr, Levy, Scheepers, &
Tily (under revision).  See http://idiom.ucsd.edu/~rlevy/papers/barr-levy-scheepers-tily-8-aug-2011-submitted.pdf.
}
\details{
\tabular{ll}{
Package: \tab simgen\cr
Type: \tab Package\cr
Version: \tab 1.5\cr
Date: \tab 2012-02-20\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
To get started, see \code{\link{createParamMx}}, \code{\link{mkDf}}, and \code{\link{mcRun}}.
}
\author{
Dale J. Barr <dale.barr@glasgow.ac.uk>
}
%\references{
%~~ Literature or other references for background information ~~
%}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%}
