\docType{package}
\name{simgen-package}
\alias{simgen}
\alias{simgen-package}
\title{Monte Carlo Simulation of Simultaneous Generalization to Subjects and Items}
\description{
  Routines for running Monte Carlo simulations testing the
  generalizability of various ways of analyzing datasets
  with crossed random effects.  This accompanies a paper by
  Barr, D. J., Levy, R., Scheepers, C., & Tily H. J. (2013),
  \href{http://dx.doi.org/10.1016/j.jml.2012.11.001}{Random effects structure for confirmatory hypothesis testing: Keep it
	maximal.}
  Journal of Memory and Language, 68, 255-278.
}
\details{
  \tabular{ll}{ Package: \tab simgen\cr Type: \tab
  Package\cr Version: \tab 1.5\cr Date: \tab 2012-02-20\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr } To
  get started, see \code{\link{createParamMx}},
  \code{\link{mkDf}}, and \code{\link{mcRun}}.
}
\author{
  Dale J. Barr <dale.barr@glasgow.ac.uk>
}

