% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{tryFit}
\alias{tryFit}
\title{Fit model with error-trapping}
\usage{
tryFit(tf.formula, tf.data, ...)
}
\arguments{
\item{tf.formula}{The model formula}

\item{tf.data}{The data frame}

\item{...}{Arguments to be passed onto the fitting function \code{\link{lmer}}}
}
\value{
A list with:
\item{value}{fitted model object from lmer}
\item{converged}{whether or not the model converged}
}
\description{
Try to fit an lmer model, catching any warnings about convergence
}
\examples{
# random intercepts only
ff <- tryFit(Resp~Cond + (1+Cond|SubjID)+(1+Cond|ItemID), mkDf())
}
\seealso{
\code{\link{tryUpdate}}
}

